Widen RNG seed to UInt64 for full Philox key entropy
Previously the seed was UInt32 and the Philox key's upper lane was
always zero, so different 64-bit seeds colliding mod 2^32 produced
identical streams. Store the seed as UInt64 and split both halves
into the key's two lanes via a new philox_key helper.

+++ src/deprecated.jl
@@ -8,5 +8,5 @@ end
 
 function Random.seed!(rng::RNG, seed::Vector{UInt32})
     Base.depwarn("seed!(rng::RNG, seed::Vector{UInt32}) is deprecated, use seed!(rng, seed::Integer) instead", :seed!)
-    Random.seed!(rng, isempty(seed) ? rand(Random.RandomDevice(), UInt32) : first(seed))
+    Random.seed!(rng, isempty(seed) ? rand(Random.RandomDevice(), UInt64) : first(seed))
 end
